requirements.txt
flask==3.1.0
gunicorn==23.0.0
mysql-connector-python==8.3.0
openpyxl==3.1.2
pytest==8.1.1
